Fix field scan finding no beacons: replace 0x2080 filter with KKM MAC gate
On real hardware the scan never listed a single beacon while the
emulator test stayed green. Root cause: the app scanned with an
Android ScanFilter on KKM's 0x2080 service UUID, but
ScanFilter.setServiceUuid only matches the advertised service-class
UUID list (AD types 0x02/0x03), and real KBeacons carry 0x2080 only
in the service DATA field (AD type 0x16). The hardware filter
therefore matched nothing in the field. The simulated peripheral
advertised the UUID in the service-class list as well, which is why
CI never caught it. KKM's own KBeaconsMgr works around the same
problem by OR-ing four filters (Eddystone FEAA, 0x2080, iBeacon and
KKM manufacturer data), which hatter's single-UUID scan API cannot
express.

Fixes:
- Scan unfiltered and classify every result with the new
  Protocol.identifyScanResult: an address with KKM's BC:57:29 OUI is
  a configurable beacon (the comparison masks the two BLE
  address-type bits, so netsim's static random FC:57:29 fleet also
  matches), a foreign MAC is remembered and skipped with one log
  line, and an opaque iOS address waits for the factory name to
  identify it.
- Replace macForAuthentication with the identity gate; listed beacons
  now always carry their auth MAC, and the now-unreachable iOS
  mac-upgrade path is gone.
- Move the app from app/Main.hs into the new library module
  KBeacon.OtaApp so the test suite can drive the real platform
  callbacks; app/Main.hs only re-exports otaAppMain.
- Make the simulated peripheral advertise like real hardware (0x2080
  as service data only), so the emulator UI test reproduces the field
  failure on the old code and proves the fix end to end.
- Add a signal-to-UI integration test group: feeding BleScanResults
  into onScanResult must list the beacon in the appView widget tree
  and request a repaint, deduplicate repeats, drop weak or foreign
  signals, and reconstruct the MAC on iOS-style results.

Prompt: okay these are the log results of kbeacon-ota-tool: [logcat
excerpt showing ble_start_scan(filter=00002080-0000-1000-8000-00805F9B34FB),
scan start/stop, and "0 device(s) found"] - however we see no beacons
show up on screen at all, investigate why and fix the bug:
https://github.com/jappeace/kbeacon-ota-tool/ also make some
integration test showing that if we get a signal the ui updates
